[OF-1494] refac: Enable warnings as errors (#591)
* [OF-1494] refac: Enable warnings as errors

Enable warnings as errors for all our projects.
Disable warnings from external projects.

* [Of-1494] refac: Suppress no PDB warning

Suppress the no PBD warning in the test runner for gtest.
We do this for the main project too.

* [OF-1494] refac: Explicitly cast types

The easy type casts for the warnings as errors ticket.

* [OF-1494] refac: Suppress memory manager warn

I'll be honest, I don't know what this does and why
it's worth warning about.
Am planning to remove the custom memory
management for exactly
this sort of reason, it's a big unknown.

* [OF-1494] refac: Cast to float in script macro

Cast to the appropriate types to avoid warnings.
This seems a bit wrong in general, we should update
ReplicatedValue so we can store this type directly.

* [OF-1494] refac: Supress DLL warning generator

We're using standard types in the generated code,
they have warnings on windows which can
be ignored around dll visibility.

* [OF-1494] refac: Tricky clang warnings

Disable a pair of tricky clang warnings in the
emscripten build. Clang can be quite
particular about its error behaviour, so we
need a bit of brute force on these ones.

* [OF-1494] refac: Handle NumElectionMessages

We wern't explicitly handling this enum.
Dosen't look like we do anything though.

* [OF-1494] refac: SignalR client warnings

Disable them, this is thirdparty.

* [OF-1494] refac: POCO warnings

Disable them, it's thirdparty.

* [OF-1494] refac: non-deprecated get/set name

on VideoPlayerSpaceComponentScriptInterface

To avoid breaking the script interface,
do not change this name, but merely
set the underlying name to be the non deprecated version.

This is a tad debtful.

* [OF-1494] refac: Remove Unused variables

Remove or flag variables that were being reported as unused.

* [OF-1494] refac: MultiPlayerTests warnings

floats, + fail macro doesnt take a msg.

* [OF-1494] refac: Brace ambiguous conditional

warnings as errors

* [OF-1494] refac: Do not underflow

This is UB

* [OF-1494] refac: Disable warnings in csharp wrapper

Because we're not messing with the generators, disable warnings.
The specific one, (although it may just be the first) :
Error CS0109: The member 'HotspotSequenceSystem.Dispose()' does not hide an accessible member. The new keyword is not required.)

* [OF-1494] refac: Deal with escape character in regex warning

Use raw strings to not deal with the escape character

* [OF-1494] test: Temporarily suppress unused variable warning

This is probably a real bug and this commit will be undone with the actual fix
in the next PR.

* [OF-1494] refac: SignalRMsgPack Warnings

A warning about not handling an enum case

* [OF-1494] refac: Temporarily deal with warnings in unused functions

These serialization functions are unused,
we're gonna remove them, but I want to do that
in a seperate PR, so just temporarily make them "work"

* fixup! [OF-1494] refac: Explicitly cast types

Reformat Test_CLI

* fixup! [OF-1494] refac: SignalRMsgPack Warnings

Type

+++ Library/src/Systems/ECommerce/ECommerce.cpp
@@ -673,8 +673,8 @@ void ProductInfoCollectionResult::OnResponse(const csp::services::ApiResponseBas
         std::vector<chs_aggregation::ShopifyProductDto>& ProductsArray = ProductCollectionResponse->GetArray();
 
         // Loop through products to count the variants, we want 1 output product per variant
-        int VariantCount = 0;
-        for (int DtoCount = 0; DtoCount < ProductsArray.size(); DtoCount++)
+        size_t VariantCount = 0;
+        for (size_t DtoCount = 0; DtoCount < ProductsArray.size(); DtoCount++)
         {
             VariantCount += ProductsArray[DtoCount].GetVariants().size();
         }
